Made changes to increase our exploration space within the problem space.  (regularizing)
changed AIPlayer to take an 'exploration' argument.  exploration = 0-1, representing ratio of random to sensible moves.

Changed SmartPlayer() to take argument "iq" as an float between 0 and 1.  the lower the number, the more likely the SmartPlayer() is to make a sub-optimal move.  Also returns a random winning move, or random opponent block if more than one are available, rather than the first one listed.