Added eval_loop().  this takes a model and runs it 1000 times against a SmartPlayer(.9).  It then writes the model information and the evaluation results to 'Experimental Dataframe.csv' in the shared drive folder.
Also changed training_loop() to save a list of opponents in a new model attribute called with ai.model.opponent_history.  This will be a list of tuples, the first element a list of opponents and the second element an integer number of epochs it trained.

in AIPlayer() I changed self.lastloss to initialize as (0,0) rather than none, so it can be called when a model is not training.  Did some other changes to support non-training models in the training_loop.