changed aiplayer to hold the model training histories and changed it to the primary object to be saved as passed around, rather than the model.
Implemented aiplayer.save() that saves the model as an .h5 and the player information as a .pickle.

Implemented AIPlayer.load(path) that loads the aiplayer and the model and puts them together.

Changed eval_loop to output an easier to read dataframe with models as columns.  still needs some work.